Improve unnecessary CurrentChat re-renders through composition ->  Create ChatInput component so that when typing a message only the input component re-renders(not the whole chat)
