+++ client/src/views/Chat/CurrentChat/ChatInput/index.js
@@ -1,0 +1,1 @@
+export * from './ChatInput';
