+++ client/src/views/Chat/CurrentChat/ChatInput/styles.module.scss
@@ -1,0 +1,7 @@
+.type-message {
+    width: 100%;
+    margin: 0;
+    &__input {
+        padding: 1rem;
+    }
+}
